mainAndPauseMenu
Added the main and pause menu into the game. The pause menu is a prefab that can just be dropped into a scene, depending on the component it may need to check whether the game is paused or not because of the Time.timescale, so there is a static bool variable called "paused" which can be used to stop or continue them. Also added a menu folder in Script for the menu scripts.